docker_example/video_streaming/saver/requirements.txt
numpy
opencv-python
kafka-python
pillow
comdutils
shapely
redis
pymysql
